refactor(data-model): capitalize single-class file names
Apply the file-naming convention to `data-model`: files that exist
entirely or very nearly entirely to export a single class get a
PascalCase name matching the class. Nine source files renamed, with
their `deno.json` export specifiers and all importers updated to
match:

- `unknown-value.ts` -> `UnknownValue.ts`
- `explicit-tag-value.ts` -> `ExplicitTagValue.ts`
- `fabric-bytes.ts` -> `FabricBytes.ts`
- `fabric-hash.ts` -> `FabricHash.ts`
- `schema-and-hash.ts` -> `SchemaAndHash.ts` (no `deno.json` export;
  internal-only)
- `problematic-value.ts` -> `ProblematicValue.ts`
- `base-reconstruction-context.ts` -> `BaseReconstructionContext.ts`
- `json-encoding-context.ts` -> `JsonEncodingContext.ts`
- `empty-reconstruction-context.ts` -> `EmptyReconstructionContext.ts`
  (file is very nearly entirely the class plus its canonical frozen
  singleton `EMPTY_RECONSTRUCTION_CONTEXT`)

Five single-class unit-test files renamed to match their subject:
`FabricBytes.test.ts`, `FabricHash.test.ts`, `SchemaAndHash.test.ts`,
`JsonEncodingContext.test.ts`, `EmptyReconstructionContext.test.ts`.

Cross-package importers in `memory`, `runner`, `runtime-client`, and
`toolshed` updated to the new capitalized export specifiers.
