+++ packages/data-model/FabricBytes.ts
