+++ packages/data-model/deno.json
@@ -21,13 +21,13 @@
   },
   "exports": {
     "./array-utils": "./array-utils.ts",
-    "./base-reconstruction-context": "./base-reconstruction-context.ts",
+    "./BaseReconstructionContext": "./BaseReconstructionContext.ts",
     "./deep-freeze": "./deep-freeze.ts",
-    "./empty-reconstruction-context": "./empty-reconstruction-context.ts",
-    "./explicit-tag-value": "./explicit-tag-value.ts",
-    "./fabric-bytes": "./fabric-bytes.ts",
+    "./EmptyReconstructionContext": "./EmptyReconstructionContext.ts",
+    "./ExplicitTagValue": "./ExplicitTagValue.ts",
+    "./FabricBytes": "./FabricBytes.ts",
     "./fabric-epoch": "./fabric-epoch.ts",
-    "./fabric-hash": "./fabric-hash.ts",
+    "./FabricHash": "./FabricHash.ts",
     "./fabric-native-instances": "./fabric-native-instances.ts",
     "./fabric-type-tags": "./fabric-type-tags.ts",
     "./fabric-value": "./fabric-value.ts",
@@ -37,13 +37,13 @@
     "./frozen-builtins": "./frozen-builtins.ts",
     "./interface": "./interface.ts",
     "./json-encoding": "./json-encoding.ts",
-    "./json-encoding-context": "./json-encoding-context.ts",
+    "./JsonEncodingContext": "./JsonEncodingContext.ts",
     "./json-type-handlers": "./json-type-handlers.ts",
     "./native-type-tags": "./native-type-tags.ts",
-    "./problematic-value": "./problematic-value.ts",
+    "./ProblematicValue": "./ProblematicValue.ts",
     "./schema-hash": "./schema-hash.ts",
     "./schema-utils": "./schema-utils.ts",
-    "./unknown-value": "./unknown-value.ts",
+    "./UnknownValue": "./UnknownValue.ts",
     "./value-debug": "./value-debug.ts",
     "./value-hash": "./value-hash.ts"
   }
