+++ packages/data-model/json-encoding.ts
@@ -1,8 +1,8 @@
 import { isInstance } from "@commonfabric/utils/types";
 import type { FabricValue } from "./fabric-value.ts";
 import type { ReconstructionContext } from "./fabric-value.ts";
-import { EmptyReconstructionContext } from "./empty-reconstruction-context.ts";
-import { JsonEncodingContext } from "./json-encoding-context.ts";
+import { EmptyReconstructionContext } from "./EmptyReconstructionContext.ts";
+import { JsonEncodingContext } from "./JsonEncodingContext.ts";
 
 /** Shared JSON encoding context. */
 const jsonEncodingContext = new JsonEncodingContext();
