+++ packages/data-model/test/EmptyReconstructionContext.test.ts
@@ -3,11 +3,11 @@ import { expect } from "@std/expect";
 import {
   EMPTY_RECONSTRUCTION_CONTEXT,
   EmptyReconstructionContext,
-} from "../empty-reconstruction-context.ts";
+} from "../EmptyReconstructionContext.ts";
 
 describe("EMPTY_RECONSTRUCTION_CONTEXT", () => {
   it("is a singleton (re-import yields the same instance)", async () => {
-    const reimported = (await import("../empty-reconstruction-context.ts"))
+    const reimported = (await import("../EmptyReconstructionContext.ts"))
       .EMPTY_RECONSTRUCTION_CONTEXT;
     expect(reimported).toBe(EMPTY_RECONSTRUCTION_CONTEXT);
   });
