+++ packages/data-model/test/SchemaAndHash.test.ts
@@ -1,8 +1,8 @@
 import { describe, it } from "@std/testing/bdd";
 import { assert } from "@std/assert";
 import { expect } from "@std/expect";
-import { FabricHash } from "../fabric-hash.ts";
-import { SchemaAndHash } from "../schema-and-hash.ts";
+import { FabricHash } from "../FabricHash.ts";
+import { SchemaAndHash } from "../SchemaAndHash.ts";
 import { toDeepFrozenSchema } from "../schema-utils.ts";
 
 // Two distinct non-empty byte arrays so tests can observe variance.
