+++ packages/memory/commit.ts
@@ -8,7 +8,7 @@ import type {
   Revision,
   Transaction,
 } from "./interface.ts";
-import { FabricHash } from "@commonfabric/data-model/fabric-hash";
+import { FabricHash } from "@commonfabric/data-model/FabricHash";
 import { assert } from "./fact.ts";
 
 export const COMMIT_LOG_TYPE = "application/commit+json" as const;
