+++ packages/toolshed/routes/blobs/blobs.test.ts
@@ -6,9 +6,9 @@ import memory from "@/routes/storage/memory/memory.index.ts";
 import { memoryServer } from "@/routes/storage/memory.ts";
 import env from "@/env.ts";
 import { Identity } from "@commonfabric/identity";
-import { FabricBytes } from "@commonfabric/data-model/fabric-bytes";
+import { FabricBytes } from "@commonfabric/data-model/FabricBytes";
 import { hashOf } from "@commonfabric/data-model/value-hash";
-import { JsonEncodingContext } from "@commonfabric/data-model/json-encoding-context";
+import { JsonEncodingContext } from "@commonfabric/data-model/JsonEncodingContext";
 import { encodeMemoryBoundary } from "@commonfabric/memory/v2";
 import { Runtime } from "@commonfabric/runner";
 import { StorageManager } from "@commonfabric/runner/storage/cache.deno";
